spliting views in entities module
