+++ entities/algorithm.py
@@ -11,6 +11,7 @@ check it can be with other room's subjects
 
 
 class ImprovementHelper():
+    # TODO: this logic should be in Directory class!!!
     bachelor_semesters = [1, 2, 3, 4, 5, 6, 7]
     master_semesters = [1, 2, 3]
     only_master_semesters = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
